requirements.txt
# Core Dependencies for WasteCycle Full-Stack Application

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Gemini AI
google-generativeai==0.3.2

# Environment Management
python-dotenv==1.0.0

# Image Processing
pillow==10.1.0

# Async File Operations
aiofiles==23.2.1